+++ src/Model/Concern/Timestampable.php
@@ -20,24 +20,44 @@ use Origin\Model\Entity;
 /**
  * Timestampable Behavior
  * adds timestamp to created and modified fields.
+ *
+ * @todo an idea on how to go about configuration
+ *
+ * $this->timestampConfig([
+ *  'created' => 'created_at',
+ *  'modified' => 'modified_at'
+ * ]);
  */
 trait Timestampable
 {
+    /**
+     * @var string
+     */
+    private $timestamp;
+
     public function initializeTimestampable()
     {
-        /**
-         * @todo Need to figure out how not to polute the method list but
-         * be able to configure each concern maybe something like
-         *
-         * $this->configConcern('Timestampable',['created'=>'created'])
-         */
         if (! isset($this->createdField)) {
             $this->createdField = 'created';
         }
         if (! isset($this->modifiedField)) {
             $this->modifiedField = 'modified';
         }
+        $this->timestamp = date('Y-m-d H:i:s');
+        $this->beforeCreate('timestambleBeforeCreate');
         $this->beforeSave('timestambleBeforeSave');
+    }
+
+    /**
+    * Before create callback
+    *
+    * @param \Origin\Model\Entity $entity
+    * @param ArrayObject $options
+    * @return void
+    */
+    protected function timestambleBeforeCreate(Entity $entity, ArrayObject $options) : void
+    {
+        $this->setTimestamp($entity, $this->createdField);
     }
 
     /**
@@ -45,28 +65,27 @@ trait Timestampable
      *
      * @param \Origin\Model\Entity $entity
      * @param ArrayObject $options
-     * @return bool must return true to continue
+     * @return void
      */
-    protected function timestambleBeforeSave(Entity $entity, ArrayObject $options) : bool
+    protected function timestambleBeforeSave(Entity $entity, ArrayObject $options) : void
     {
-        $timestamp = date('Y-m-d H:i:s');
-        $primaryKey = $this->primaryKey;
-
-        if (empty($entity->$primaryKey)) {
-            $this->setTimestamp($entity, $this->createdField, $timestamp);
-        }
-        $this->setTimestamp($entity, $this->modifiedField, $timestamp);
-        
-        return true;
+        $this->setTimestamp($entity, $this->modifiedField);
     }
 
-    private function setTimestamp(Entity $entity, string $field, string $timestamp)
+    /**
+     * Sets the timestamp
+     *
+     * @param \Origin\Model\Entity $entity
+     * @param string $field
+     * @return void
+     */
+    private function setTimestamp(Entity $entity, string $field) : void
     {
         if (! $this->hasField($field)) {
             return;
         }
         if (empty($entity->$field) or ! in_array($field, $entity->modified())) {
-            $entity->set($field, $timestamp);
+            $entity->set($field, $this->timestamp);
         }
     }
 }